# docs/openapi.yaml
openapi: 3.0.3
info:
  title: クローゼット管理アプリ API
  version: 1.1.0
  description: |
    クローゼット管理アプリのバックエンドAPI仕様。
    - LLM: Google Gemini 2.5 Flash
    - 天気: Open-Meteo
    - 決済: Stripe Checkout / Webhook
    - 認証・DB・Storage: Supabase
...
servers:
  - url: http://localhost:8000
    description: ローカル開発
  - url: https://api.example.com
    description: 本番（努力目標）
security:
  - BearerAuth: []

# ================================================================
# 共通レスポンス定義
# ================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Authが発行するJWT
  responses:
    Unauthorized:
      description: 未認証（JWTが無効・期限切れ）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "認証が必要です"
            code: "unauthorized"
    Forbidden:
      description: 認可エラー（他ユーザーのリソースへのアクセス）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "アクセス権限がありません"
            code: "forbidden"
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "リソースが見つかりません"
            code: "not_found"
...
    TooManyRequests:
      description: レート制限超過
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "リクエスト回数の上限に達しました。しばらく待ってから再試行してください"
            code: "rate_limit_exceeded"
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: エラー内容
        code:
          type: string
          nullable: true
          description: アプリケーション側のエラーコード

    # ===== Auth =====
...
    ClothingCategory:
      type: string
      enum: [tops, bottoms, outer, shoes, bag, accessory]
      description: |
        服のカテゴリ。
        - tops: トップス
        - bottoms: ボトムス
        - outer: アウター
        - shoes: 靴
        - bag: バッグ
        - accessory: アクセサリー
    ClothingPattern:
      type: string
      enum: [solid, stripe, check, dot, floral, other]
      description: |
        服の柄。
        - solid: 無地
        - stripe: ストライプ
        - check: チェック
        - dot: ドット
        - floral: 花柄
        - other: その他
    SeasonTag:
      type: string
      enum: [spring, summer, autumn, winter, all]
    TpoTag:
      type: string
      enum: [casual, formal, ceremony, leisure, business]
    ClothingItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
          example: 白シャツ
        category:
          $ref: "#/components/schemas/ClothingCategory"
        color:
          type: string
          example: white
        pattern:
          allOf:
            - $ref: "#/components/schemas/ClothingPattern"
          nullable: true
        size:
          type: string
          example: M
          nullable: true
        season:
          type: array
          items:
            $ref: "#/components/schemas/SeasonTag"
        tpo_tags:
          type: array
          items:
            $ref: "#/components/schemas/TpoTag"
          description: このアイテムが適したTPOタグ（clothes_tpo テーブルに対応）
        image_url:
          type: string
          format: uri
          description: Supabase Storage の署名付きURL
          nullable: true
        thumbnail_url:
          type: string
          format: uri
          nullable: true
        memo:
          type: string
          maxLength: 200
          nullable: true
        is_favorite:
          type: boolean
          default: false
        wear_count:
          type: integer
          default: 0
          description: 着用回数（MVPでは記録のみ、学習には未使用）
        last_worn_at:
          type: string
          format: date-time
          nullable: true
        attributes:
          type: object
          description: LLMが画像から抽出した追加属性（素材・スタイルなど）
          additionalProperties: true
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    OutfitItem:
      type: object
      description: コーデを構成する1アイテムと、そのコーデ内での役割
      properties:
        clothes_id:
          type: string
          format: uuid
        role:
          allOf:
            - $ref: "#/components/schemas/ClothingCategory"
          description: このコーデ内でのアイテムの役割
        display_order:
          type: integer
          description: 表示順
        clothing_item:
          allOf:
            - $ref: "#/components/schemas/ClothingItem"
          description: 服の詳細情報（outfit_items JOIN clothes）
    Outfit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        tpo:
          $ref: "#/components/schemas/TpoTag"
        region_code:
          type: string
          example: "19_02"
          description: 提案時に使用した地域コード
        weather_summary:
          type: string
          description: 提案時の天気スナップショット
          nullable: true
        weather_temp_max:
          type: number
          nullable: true
        weather_temp_min:
          type: number
          nullable: true
        comment:
          type: string
          description: LLMのおすすめポイント（200字以内）
          nullable: true
        coordinate_image_url:
          type: string
          format: uri
          description: 生成または保存済みのコーデ画像URL
          nullable: true
        is_favorite:
          type: boolean
          default: false
        source:
          type: string
          enum: [llm, manual]
          description: llm = LLM提案, manual = ユーザーが手動で作成
        items:
          type: array
          items:
            $ref: "#/components/schemas/OutfitItem"
        created_at:
          type: string
          format: date-time
    OutfitsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Outfit"
        total:
          type: integer
    OutfitUpdateRequest:
      type: object
      description: お気に入りトグルなど、提案済みコーデの属性更新
      properties:
        is_favorite:
          type: boolean

    # ===== コーデ提案リクエスト／レスポンス =====
    CoordinateSuggestRequest:
      type: object
      required: [tpo, date]
      properties:
        tpo:
          $ref: "#/components/schemas/TpoTag"
        date:
          type: string
          format: date
          description: 着用日（天気取得に使用）
        region_code:
          type: string
          example: "19_02"
          description: |
            提案で使用する地域コード。
            省略時はユーザーの default_region_code を使用。
            旅行先・帰省先など別地域での提案を受けたい場合に指定する。
        clothing_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 提案に必ず含める服のID（指定アイテムは結果に必ず入る。残りの枠は手持ちから自動補完）
        exclude_clothing_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 提案から除外する服のID（「別の提案」時に前回のアイテムを除外するなど）

    # LLM提案アイテム（POST /outfits/suggest 用）。
    # 手持ち服はそのまま clothing_item に解決し、手持ちにない補完アイテムは
    # clothing_item=null（name/role/color/pattern のみ）で返すハイブリッド形式。
    SuggestOutfitItem:
      type: object
      description: LLMが提案したコーデの1アイテム
      properties:
        name:
          type: string
          description: アイテム名（手持ち／補完提案いずれも入る）
        role:
          type: string
          description: このコーデ内でのアイテムの役割（tops / bottoms / outer / onepiece / shoes / bag / accessory）
        color:
          type: string
          nullable: true
        pattern:
          type: string
          nullable: true
        display_order:
          type: integer
          description: 表示順
        clothing_item:
          allOf:
            - $ref: "#/components/schemas/ClothingItem"
          nullable: true
          description: 手持ち服に一致した場合の服詳細。手持ちにない補完提案の場合は null
    SuggestOutfit:
      type: object
      description: LLMが提案したコーデ（現状はDB保存せず、id/created_atは一時生成）
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        tpo:
          $ref: "#/components/schemas/TpoTag"
        comment:
          type: string
          description: LLMのおすすめポイント
          nullable: true
        is_favorite:
          type: boolean
          default: false
        items:
          type: array
          items:
            $ref: "#/components/schemas/SuggestOutfitItem"
        created_at:
          type: string
          format: date-time
    CoordinateSuggestResponse:
      type: object
      description: |
        コーデ提案レスポンス。クローゼットを参照し LLM がコーデを生成する。
        現状はテキスト表示を優先し DB 保存しないため、outfits のみを返す
        （旧 weather_summary / region_used / cached は廃止）。
      properties:
        outfits:
          type: array
          items:
            $ref: "#/components/schemas/SuggestOutfit"
          description: 提案されたコーデ

    # ===== 天気 =====
...
paths:
...
  /api/v1/outfits/suggest:
    post:
      tags: [Outfit]
      operationId: suggestOutfits
      summary: LLMによるコーデ提案
      description: |
        登録済みの服（クローゼット） × TPO × 天気情報を LLM に渡してコーデを提案する。
        手持ち服を優先し、手持ちで埋まらないカテゴリは LLM が補完アイテムを提案する
        （ハイブリッド）。手持ちに一致したアイテムは clothing_item に解決して返す。

        現状はテキスト提案の表示を優先するため DB 保存は行わない（履歴化は後続対応）。
        天気キャッシュ（Redis 30分）は提案入力として利用する。

        レート制限: 1ユーザー1分10回 / 1日の上限はプランによる
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoordinateSuggestRequest"
      responses:
        "200":
          description: コーデ提案結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoordinateSuggestResponse"
        "400":
          description: 不正な region_code、またはユーザーの default_region_code 未設定
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/outfits:
    get:
      tags: [Outfit]
      operationId: listOutfits
      summary: 提案コーデ履歴一覧
      description: ログインユーザーの過去の提案コーデ一覧を返す。
      parameters:
        - name: is_favorite
          in: query
          schema:
            type: boolean
          description: true の場合、お気に入りのみ返す
        - name: tpo
          in: query
          schema:
            $ref: "#/components/schemas/TpoTag"
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: 取得開始日（created_at）
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: 取得終了日（created_at）
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: コーデ一覧
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutfitsListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/outfits/{outfit_id}:
    parameters:
      - name: outfit_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Outfit]
      operationId: getOutfitById
      summary: コーデ詳細取得
      responses:
        "200":
          description: コーデ詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outfit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Outfit]
      operationId: updateOutfitById
      summary: コーデの属性更新（お気に入りトグルなど）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutfitUpdateRequest"
      responses:
        "200":
          description: 更新後のコーデ情報
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outfit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ================================================================
  # 天気
  # ================================================================
...